Add options for exponential backoff with task autoretry (#4101)
* Add options for exponential backoff with task autoretry

* Add test for exponential backoff

* closer to a fixed test

* Move autoretry backoff functionality inside run wrapper

* Add a test for jitter

* Correct for semantics of `random.randrange()`

`random.randrange()` treats the argument it receives as just *outside* the
bound of possible return values. For example, if you call
`random.randrange(2)`, you might get 0 or 1, but you'll never get 2.
Since we want to allow the `retry_jitter` parameter to occasionally apply no
jitter at all, we need to add one to the value we pass to `randrange()`,
so that there's a chance that we receive that original value back.

* Put side_effect on patch lines

* Fix flake8

* Add celery.utils.time.get_exponential_backoff_interval

* Use exponential backoff calculation from utils in task

* Update docs around retry_jitter

* Remove unnecessary random.choice patching

* Update task auto-retry documentation

* PEP8: remove unused import

* PEP8: remove trailing whitespace

* PEP8: Fix E123 warning